// precotex_angular/src/app/models/Tintoreria/Capacidades.ts
export class Capacidades{
    Fec_Creacion: string = '';
    Cod_Familia: string = '';
    Cod_Tela: string=  '';
    Cod_Cliente: string = '';
    Nom_Cliente: string = '';
    Tip_Ancho: string = '';
    Des_Tip_Ancho: string = '';
    Cod_Gama: string = '';
    Des_Gama: string = '';
    Eco_Master: string = '';
    IMaster: string = '';
    TRD: string = '';
    ATYC: string = '';
    MS: string = '';
    Obs_Eco_Master: string = '';
    Obs_IMaster: string = '';
    Obs_TRD: string = '';
    Obs_ATYC: string = '';
    Obs_MS: string = '';
    Fec_Reg_Ini: string = '';
    Fec_Reg_Fin: string = '';
}